# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GDT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exerciciosGetAll/{pagina}": {
            "get": {
                "tags": [
                    "Exercícios"
                ],
                "summary": "Retorna lista de exercícios",
                "operationId": "4693b05af9c850e64230cfc383d12413",
                "parameters": [
                    {
                        "name": "página",
                        "in": "path",
                        "description": "Página que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de exercícios"
                    }
                }
            }
        },
        "/api/exercicios/{id}": {
            "get": {
                "tags": [
                    "Exercícios"
                ],
                "summary": "Retorna um exercicio pelo id",
                "operationId": "4320a4589ba49880cf59619f6a7fc0a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do exercício que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exercício"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Exercícios"
                ],
                "summary": "Exclui um exercício",
                "operationId": "abf6a358019f3ad818a2367a2cdef771",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do exercício que deseja excluir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/exercicios/": {
            "put": {
                "tags": [
                    "Exercícios"
                ],
                "summary": "Atualiza um exercício",
                "operationId": "8bbf96004f657a3b81dfdb1bbc1266e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "nome",
                                    "is_aparelho"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "Supino reto"
                                    },
                                    "is_aparelho": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercício atualizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Exercícios"
                ],
                "summary": "Cadastra um exercício",
                "operationId": "170ae079ef156cd84728cb16d384f560",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "is_aparelho"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Supino reto"
                                    },
                                    "is_aparelho": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exercício criado"
                    }
                }
            }
        },
        "/api/treinosGetAll/{pagina}": {
            "get": {
                "tags": [
                    "Treinos"
                ],
                "summary": "Retorna lista de treinos",
                "operationId": "08e930a92403b4b1a2fb97df95f8922f",
                "parameters": [
                    {
                        "name": "página",
                        "in": "path",
                        "description": "Página que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de treinos"
                    }
                }
            }
        },
        "/api/treinos/{id}": {
            "get": {
                "tags": [
                    "Treinos"
                ],
                "summary": "Retorna um exercicio pelo id",
                "operationId": "33cb78dec323473267768bcfcf79db77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do treino que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Treino"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Treinos"
                ],
                "summary": "Exclui um treino",
                "operationId": "e7252d6bfcfb061aade51af83592a384",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do treino que deseja excluir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/treinos/": {
            "put": {
                "tags": [
                    "Treinos"
                ],
                "summary": "Atualiza um treino",
                "operationId": "c2fe531f89fdfcb2feb9c24eb3ea5811",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "grupamento_muscular",
                                    "foco"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grupamento_muscular": {
                                        "type": "string",
                                        "example": "Supino reto"
                                    },
                                    "foco": {
                                        "type": "string",
                                        "example": "Hipertrofia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Treino atualizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Treinos"
                ],
                "summary": "Cadastra um treino",
                "operationId": "8e8a7c149c6c195aba9ccd336b43223b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grupamento_muscular",
                                    "foco"
                                ],
                                "properties": {
                                    "grupamento_muscular": {
                                        "type": "string",
                                        "example": "Supino reto"
                                    },
                                    "foco": {
                                        "type": "string",
                                        "example": "Hipertrofia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Treino atualizado"
                    }
                }
            }
        }
    },
...
}